platform_api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
asyncpg==0.29.0
redis==5.0.7
python-multipart==0.0.9
requests==2.32.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
